// packages/react-ui/src/layout/types.ts
export type BoxElement =
	| 'article'
	| 'aside'
	| 'div'
	| 'footer'
	| 'header'
	| 'li'
	| 'nav'
	| 'ol'
	| 'section'
	| 'span'
	| 'ul';

export type BoxAlignContent =
	| 'baseline'
	| 'center'
	| 'end'
	| 'flex-end'
	| 'flex-start'
	| 'normal'
	| 'safe'
	| 'space-around'
	| 'space-between'
	| 'space-evenly'
	| 'start'
	| 'stretch'
	| 'unsafe';

export type BoxAlignItems =
	| 'baseline'
	| 'center'
	| 'end'
	| 'flex-end'
	| 'flex-start'
	| 'normal'
	| 'safe'
	| 'self-end'
	| 'self-start'
	| 'start'
	| 'stretch'
	| 'unsafe';

export type BoxAlignSelf =
	| 'auto'
	| 'baseline'
	| 'center'
	| 'flex-end'
	| 'flex-start'
	| 'normal'
	| 'safe'
	| 'self-end'
	| 'self-start'
	| 'stretch'
	| 'unsafe';

export type BoxDirection = 'column-reverse' | 'column' | 'row-reverse' | 'row';

export type BoxJustifyContent =
	| 'baseline'
	| 'center'
	| 'end'
	| 'flex-end'
	| 'flex-start'
	| 'left'
	| 'normal'
	| 'right'
	| 'safe'
	| 'space-around'
	| 'space-between'
	| 'space-evenly'
	| 'start'
	| 'unsafe';